newsletter/requirements.txt
anthropic>=0.40.0
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.120.0
python-dotenv>=1.0.0
markdown>=3.6
